getAccountById() in rest and service + tests

+++ src/main/java/com/benmorant/katas/bankingapp/similibanq/entity/CurrentAccount.java
@@ -11,7 +11,7 @@ public class CurrentAccount extends Account {
   public CurrentAccount() {}
 
   public CurrentAccount(
-      Integer idAccount, double balance, Customer customer, List<BankOperation> bankOperations) {
+      Long idAccount, double balance, Customer customer, List<BankOperation> bankOperations) {
     super(idAccount, balance, customer, bankOperations);
   }
 }
